pyproject.toml
[project]
name = "source-pad"
version = "0.1.0"
description = "Index your code, ask questions. Works with local LLMs."
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
    "llama-index-core>=0.12.0",
    "llama-index-vector-stores-chroma>=0.4.0",
    "llama-index-embeddings-ollama>=0.5.0",
    "llama-index-llms-ollama>=0.5.0",
    "llama-index-llms-openai-like>=0.3.0",
    "llama-index-readers-github>=0.5.0",
    "chromadb>=0.6.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",
]

[project.scripts]
source-pad = "source_pad.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/source_pad"]
